add class for FRACTAL dataset - add configs for concat experiments - add config for FRACTAL and ECLAIR datasets

+++ ml3d/datasets/delft_sum.py
@@ -180,7 +180,7 @@ class SUMDataset(BaseDataset):
         make_dir(path)
 
         pred = results["predict_labels"]
-        pred = np.array(pred)
+        pred = np.array(pred).astype(np.int32)
 
         for ign in cfg.ignored_label_inds:
             pred[pred >= ign] += 1
